Add files via upload
Here is a the Slack script that I use in my environment.
I recreated the Slack install script without the jq dependency.  It will update Slack if it's not on the latest version and it fixes the ownership issue.  
I also added option to stop Slack from running if "kill" is passed in parameter 4

Special thanks to [~owen.pragel] and [~anverhousseini] for their scripts